+++ backend/src/baserow/contrib/database/api/fields/errors.py
@@ -1,5 +1,6 @@
 from rest_framework.status import HTTP_400_BAD_REQUEST, HTTP_404_NOT_FOUND
 
+from baserow.contrib.database.fields.handler import RESERVED_BASEROW_FIELD_NAMES
 
 ERROR_FIELD_DOES_NOT_EXIST = (
     "ERROR_FIELD_DOES_NOT_EXIST",
@@ -50,6 +51,11 @@ ERROR_FIELD_WITH_SAME_NAME_ALREADY_EXISTS = (
 ERROR_RESERVED_BASEROW_FIELD_NAME = (
     "ERROR_RESERVED_BASEROW_FIELD_NAME",
     HTTP_400_BAD_REQUEST,
-    "The provided field name is a reserved Baserow field name and cannot be used for a"
-    "user created field, please choose different field name.",
+    f"The field names {','.join(RESERVED_BASEROW_FIELD_NAMES)} are reserved and cannot "
+    f"and cannot be used for a user created field, please choose different field name.",
+)
+ERROR_INVALID_BASEROW_FIELD_NAME = (
+    "ERROR_INVALID_BASEROW_FIELD_NAME",
+    HTTP_400_BAD_REQUEST,
+    "Fields must not be blank or only consist of whitespace.",
 )