+++ backend/src/baserow/contrib/database/api/fields/views.py
@@ -28,6 +28,7 @@ from baserow.contrib.database.api.fields.errors import (
     ERROR_MAX_FIELD_COUNT_EXCEEDED,
     ERROR_RESERVED_BASEROW_FIELD_NAME,
     ERROR_FIELD_WITH_SAME_NAME_ALREADY_EXISTS,
+    ERROR_INVALID_BASEROW_FIELD_NAME,
 )
 from baserow.contrib.database.fields.exceptions import (
     CannotDeletePrimaryField,
@@ -36,6 +37,7 @@ from baserow.contrib.database.fields.exceptions import (
     MaxFieldLimitExceeded,
     ReservedBaserowFieldNameException,
     FieldWithSameNameAlreadyExists,
+    InvalidBaserowFieldName,
 )
 from baserow.contrib.database.fields.models import Field
 from baserow.contrib.database.fields.handler import FieldHandler
@@ -146,6 +148,7 @@ class FieldsView(APIView):
                     "ERROR_MAX_FIELD_COUNT_EXCEEDED",
                     "ERROR_RESERVED_BASEROW_FIELD_NAME",
                     "ERROR_FIELD_WITH_SAME_NAME_ALREADY_EXISTS",
+                    "ERROR_INVALID_BASEROW_FIELD_NAME",
                 ]
             ),
             401: get_error_schema(["ERROR_NO_PERMISSION_TO_TABLE"]),
@@ -164,6 +167,7 @@ class FieldsView(APIView):
             NoPermissionToTable: ERROR_NO_PERMISSION_TO_TABLE,
             FieldWithSameNameAlreadyExists: ERROR_FIELD_WITH_SAME_NAME_ALREADY_EXISTS,
             ReservedBaserowFieldNameException: ERROR_RESERVED_BASEROW_FIELD_NAME,
+            InvalidBaserowFieldName: ERROR_INVALID_BASEROW_FIELD_NAME,
         }
     )
     def post(self, request, data, table_id):
@@ -263,6 +267,7 @@ class FieldView(APIView):
                     "ERROR_REQUEST_BODY_VALIDATION",
                     "ERROR_RESERVED_BASEROW_FIELD_NAME",
                     "ERROR_FIELD_WITH_SAME_NAME_ALREADY_EXISTS",
+                    "ERROR_INVALID_BASEROW_FIELD_NAME",
                 ]
             ),
             404: get_error_schema(["ERROR_FIELD_DOES_NOT_EXIST"]),
@@ -276,6 +281,7 @@ class FieldView(APIView):
             CannotChangeFieldType: ERROR_CANNOT_CHANGE_FIELD_TYPE,
             FieldWithSameNameAlreadyExists: ERROR_FIELD_WITH_SAME_NAME_ALREADY_EXISTS,
             ReservedBaserowFieldNameException: ERROR_RESERVED_BASEROW_FIELD_NAME,
+            InvalidBaserowFieldName: ERROR_INVALID_BASEROW_FIELD_NAME,
         }
     )
     def patch(self, request, field_id):