+++ backend/tests/baserow/contrib/database/api/tables/test_table_views.py
@@ -42,6 +42,17 @@ def test_list_tables(api_client, data_fixture):
     assert response.status_code == HTTP_404_NOT_FOUND
     assert response.json()["error"] == "ERROR_APPLICATION_DOES_NOT_EXIST"
 
+    response = api_client.delete(
+        reverse("api:groups:item", kwargs={"group_id": database.group.id}),
+        HTTP_AUTHORIZATION=f"JWT {token}",
+    )
+    assert response.status_code == HTTP_204_NO_CONTENT
+
+    url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
+    response = api_client.get(url, **{"HTTP_AUTHORIZATION": f"JWT {token}"})
+    assert response.status_code == HTTP_404_NOT_FOUND
+    assert response.json()["error"] == "ERROR_APPLICATION_DOES_NOT_EXIST"
+
 
 @pytest.mark.django_db
 def test_create_table(api_client, data_fixture):
@@ -286,8 +297,7 @@ def test_create_table_with_data(api_client, data_fixture):
     )
     response_json = response.json()
     assert response.status_code == HTTP_400_BAD_REQUEST
-    assert response_json["error"] == "ERROR_INVALID_INITIAL_TABLE_DATA"
-    assert "reserved" in response_json["detail"]
+    assert response_json["error"] == "ERROR_RESERVED_BASEROW_FIELD_NAME"
 
     url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
     response = api_client.post(
@@ -307,7 +317,7 @@ def test_create_table_with_data(api_client, data_fixture):
     )
     response_json = response.json()
     assert response.status_code == HTTP_400_BAD_REQUEST
-    assert response_json["error"] == "ERROR_INVALID_INITIAL_TABLE_DATA"
+    assert response_json["error"] == "ERROR_INITIAL_TABLE_DATA_HAS_DUPLICATE_NAMES"
     assert "unique" in response_json["detail"]
 
     url = reverse("api:database:tables:list", kwargs={"database_id": database.id})
@@ -327,7 +337,7 @@ def test_create_table_with_data(api_client, data_fixture):
     )
     response_json = response.json()
     assert response.status_code == HTTP_400_BAD_REQUEST
-    assert response_json["error"] == "ERROR_INVALID_INITIAL_TABLE_DATA"
+    assert response_json["error"] == "ERROR_INVALID_BASEROW_FIELD_NAME"
     assert "blank" in response_json["detail"]
 
     url = reverse("api:database:tables:list", kwargs={"database_id": database.id})