+++ web-frontend/modules/database/components/field/FieldForm.vue
@@ -11,14 +11,25 @@
           placeholder="Name"
           @blur="$v.values.name.$touch()"
         />
-        <div v-if="!$v.values.name.required" class="error">
+        <div
+          v-if="$v.values.name.$dirty && !$v.values.name.required"
+          class="error"
+        >
           This field is required.
         </div>
-        <div v-else-if="!$v.values.name.mustHaveUniqueFieldName" class="error">
+        <div
+          v-else-if="
+            $v.values.name.$dirty && !$v.values.name.mustHaveUniqueFieldName
+          "
+          class="error"
+        >
           A field with this name already exists.
         </div>
         <div
-          v-else-if="!$v.values.name.mustNotClashWithReservedName"
+          v-else-if="
+            $v.values.name.$dirty &&
+            !$v.values.name.mustNotClashWithReservedName
+          "
           class="error"
         >
           This field name is not allowed.
@@ -63,6 +74,7 @@ import { required } from 'vuelidate/lib/validators'
 
 import form from '@baserow/modules/core/mixins/form'
 import { mapGetters } from 'vuex'
+import { RESERVED_BASEROW_FIELD_NAMES } from '@baserow/modules/database/utils/constants'
 
 // @TODO focus form on open
 export default {
@@ -123,7 +135,7 @@ export default {
       return !fields.map((f) => f.name).includes(param.trim())
     },
     mustNotClashWithReservedName(param) {
-      return !['id', 'order'].includes(param.trim())
+      return !RESERVED_BASEROW_FIELD_NAMES.includes(param.trim())
     },
     getFormComponent(type) {
       return this.$registry.get('field', type).getFormComponent()