+++ web-frontend/modules/database/components/table/TableXMLImporter.vue
@@ -154,15 +154,13 @@ export default {
         return
       }
 
-      try {
-        this.values.data = JSON.stringify(xmlData)
-        this.error = ''
-        this.preview = this.getPreview(xmlData, hasHeader)
-      } catch (error) {
-        this.values.data = ''
-        this.error = error.message
-        this.preview = {}
-      }
+      const dataWithHeader = this.ensureHeaderExistsAndIsValid(
+        xmlData,
+        hasHeader
+      )
+      this.values.data = JSON.stringify(dataWithHeader)
+      this.error = ''
+      this.preview = this.getPreview(dataWithHeader)
     },
   },
 }