+++ changelog.md
@@ -3,8 +3,15 @@
 ## Unreleased
 
 * Made it possible to list table field meta-data with a token.
+* The email field's validation is now consistent and much more permissive allowing most 
+  values which look like email addresses.
+* Add trash where deleted apps, groups, tables, fields and rows can be restored 
+  deletion.
 * Fix the create group invite endpoint failing when no message provided.
 * Single select options can now be ordered by drag and drop. 
+* Added before and after date filters.
+* Support building Baserow out of the box on Ubuntu by lowering the required docker
+  version to build Baserow down to 19.03.
 * Disallow duplicate field names in the same table, blank field names or field names
   called 'order' and 'id'. Existing invalid field names will be fixed automatically. 
 * Add user_field_names GET flag to various endpoints which switches the API to work